Attempt to fix the module reloading issues
When I was testing omnibot before, I was running in to issues with code
reloading not working. I've added a call to
`importlib.invalidate_caches()` in the `ModuleLoader.load` method, as
well as a gen-0 garbage collector call upon module unload and the module
is dereferenced.

I should really consider just handing out weak references instead.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>