Add data directory option to server and module config
Modules can have their own data directory, which is lazily handed out
and created upon request by the implementing module. This is useful
because it separates the data from the code and gives everything a place
to live.

Additionally, keyboard interrupts are caught now and the program
attempts to gracefully shut down when that is caught (i.e. unload
modules).

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>